changed setup, now functions are organized by type in separate directories
