+++ include/finmath/Helper/helper.h
@@ -3,5 +3,6 @@
 
 double normal_cdf(double x);
 double normal_pdf(double x);
+double combinations(int n, int k);
 
 #endif
