+++ include/finmath/Interest and Annuities/compound_interest.h
