+++ include/finmath/finmath.h
@@ -1,9 +1,9 @@
 #ifndef FINMATH_H
 #define FINMATH_H
 
-#include "finmath/helper.h"
-#include "finmath/compound_interest.h"
-#include "finmath/black_scholes.h"
+#include "finmath/Helper/helper.h"
+#include "finmath/Interest and Annuities/compound_interest.h"
+#include "finmath/Options Pricing/black_scholes.h"
 
 // Include other headers as needed
