+++ src/cpp/Helper/helper.cpp
@@ -1,5 +1,5 @@
 #include <cmath>
-#include "finmath/helper.h"
+#include "finmath/Helper/helper.h"
 
 
 // Standard normal cumulative distribution function
