+++ src/cpp/Option Pricing/black_scholes.cpp
@@ -1,6 +1,6 @@
 #include <cmath> 
-#include "finmath/black_scholes.h"
-#include "finmath/helper.h"
+#include "finmath/Options Pricing/black_scholes.h"
+#include "finmath/Helper/helper.h"
 
 double black_scholes(OptionType type, double strike, double price, double time, double rate, double volatility){
     double d1 = (std::log(price / strike) + (1/2) + ((rate + std::pow(volatility, 2)) * time)) / (volatility * std::sqrt(time));
